Add Installer abstraction
