+++ library/binary-core-api/api/binary-core-api.api
@@ -3,3 +3,22 @@ public final class io/matthewnelson/kmp/tor/binary/core/api/Exceptions {
 	public static final fun wrapIOException (Ljava/lang/Throwable;Lkotlin/jvm/functions/Function0;)Ljava/io/IOException;
 }
 
+public abstract class io/matthewnelson/kmp/tor/binary/core/api/Installer {
+	public final field installationDir Ljava/lang/String;
+	public fun <init> (Ljava/lang/String;)V
+	public abstract fun install ()Lio/matthewnelson/kmp/tor/binary/core/api/Installer$Paths;
+}
+
+public abstract class io/matthewnelson/kmp/tor/binary/core/api/Installer$Paths {
+}
+
+public final class io/matthewnelson/kmp/tor/binary/core/api/Installer$Paths$Tor : io/matthewnelson/kmp/tor/binary/core/api/Installer$Paths {
+	public final field geoip Ljava/lang/String;
+	public final field geoip6 Ljava/lang/String;
+	public final field tor Ljava/lang/String;
+	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
+	public fun equals (Ljava/lang/Object;)Z
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
