+++ library/binary/api/android/binary.api
@@ -1,16 +1,6 @@
-public final class io/matthewnelson/kmp/tor/binary/KmpTorBinary {
-	public final field installationDir Ljava/lang/String;
+public final class io/matthewnelson/kmp/tor/binary/KmpTorBinary : io/matthewnelson/kmp/tor/binary/core/api/Installer {
 	public fun <init> (Ljava/lang/String;)V
-	public final fun install ()Lio/matthewnelson/kmp/tor/binary/KmpTorBinary$Paths;
-}
-
-public final class io/matthewnelson/kmp/tor/binary/KmpTorBinary$Paths {
-	public final field geoip Ljava/lang/String;
-	public final field geoip6 Ljava/lang/String;
-	public final field tor Ljava/lang/String;
-	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
-	public fun equals (Ljava/lang/Object;)Z
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
+	public fun install ()Lio/matthewnelson/kmp/tor/binary/core/api/Installer$Paths$Tor;
+	public synthetic fun install ()Lio/matthewnelson/kmp/tor/binary/core/api/Installer$Paths;
 }
